# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  request.Address:
    properties:
      area:
        type: string
      city:
        type: string
      house:
        type: string
      is_default:
        type: boolean
      land_mark:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      phone_number:
        maxLength: 10
        minLength: 10
        type: string
      pincode:
        type: integer
    required:
    - house
    - land_mark
    - name
    - phone_number
    - pincode
    type: object
...
  request.EditAddress:
    properties:
      address_id:
        type: integer
      area:
        type: string
      city:
        type: string
      house:
        type: string
      is_default:
        type: boolean
      land_mark:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      phone_number:
        maxLength: 10
        minLength: 10
        type: string
      pincode:
        type: integer
    required:
    - address_id
    - house
    - land_mark
    - name
    - phone_number
    - pincode
    type: object
...
  request.EditUser:
    properties:
      age:
        type: integer
      confirm_password:
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      password:
        type: string
      phone:
        maxLength: 10
        minLength: 10
        type: string
      user_name:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - age
    - email
    - first_name
    - last_name
    - phone
    - user_name
    type: object
...
  request.UpdateStock:
    properties:
      qty_to_add:
        type: integer
      sku:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: okeke98@gmail.com
    name: For API Support
  description: 'Backend API built with Golang using Clean Code architecture. \nGithub:
    [https://github.com/okekefrancis112/ecommerce-gin].'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-commerce Application Backend API
paths:
  /account:
    get:
      description: API for user to get all user details
      operationId: GetProfile
      responses:
        "200":
          description: Successfully retrieved user details
        "500":
          description: Failed to retrieve user details
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get User Profile (User)
      tags:
      - User Profile
    put:
      description: API for user to edit user details
      operationId: UpdateProfile
      parameters:
      - description: User details input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.EditUser'
      responses:
        "200":
          description: Successfully profile updated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to update profile
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Edit profile (User)
      tags:
      - User Profile
  /account/address:
    get:
      description: API for user to get all user addresses
      operationId: GetAllAddresses
      responses:
        "200":
          description: successfully retrieved all user addresses
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: failed to show user addresses
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all addresses (User)
      tags:
      - User Profile
    post:
      description: API for user to add a new address
      operationId: SaveAddress
      parameters:
      - description: Address input
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Address'
      responses:
        "200":
          description: Successfully address added
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to save address
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add a new address (User)
      tags:
      - User Profile
    put:
      description: API for user to update user address
      operationId: UpdateAddress
      parameters:
      - description: Address input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.EditAddress'
      responses:
        "200":
          description: successfully addresses updated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: can't update the address
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update address (User)
      tags:
      - User Profile
...
  /account/wishlist:
    get:
      operationId: GetWishList
      responses:
        "200":
          description: Successfully retrieved all product items in th wish list
        "500":
          description: Failed to retrieve product items from the wish list
      security:
      - BearerAuth: []
      summary: Get whish list product items (User)
      tags:
      - User Profile
  /account/wishlist/{product_item_id}:
    delete:
      operationId: RemoveFromWishList
      parameters:
      - description: Product Item ID
        in: path
        name: product_item_id
        required: true
        type: integer
      responses:
        "200":
          description: successfully removed product item from wishlist
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove from whish list (User)
      tags:
      - User Profile
    post:
      operationId: SaveToWishList
      parameters:
      - description: Product Item ID
        in: path
        name: product_item_id
        required: true
        type: integer
      responses:
        "200":
          description: Successfully product items added to whish list
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Product item already exist on wish list
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to add product item to wishlist
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add to whish list (User)
      tags:
      - User Profile
...
  /admin/stocks:
    get:
      description: API for admin to get all stocks
      operationId: GetAllStocks
      parameters:
      - description: Page Number
        in: query
        name: page_number
        type: integer
      - description: Count
        in: query
        name: count
        type: integer
      responses:
        "200":
          description: Successfully found all stocks
          schema:
            $ref: '#/definitions/response.Response'
        "204":
          description: No stocks found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to Get all stocks
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all stocks (Admin)
      tags:
      - Admin Stock
    patch:
      description: API for admin to update stock details
      operationId: UpdateStock
      parameters:
      - description: Update stock details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UpdateStock'
      responses:
        "200":
          description: Successfully updated sock
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to bind input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to update stock
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update stocks (Admin)
      tags:
      - Admin Stock
...
securityDefinitions:
  BearerAuth:
    description: 'Add prefix of Bearer before  token Ex: "Bearer token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"